+++ R/ifcb_download_dashboard_metadata.R
@@ -7,12 +7,14 @@
 #'
 #' @return A data frame containing the exported metadata.
 #' @examples
-#' \dontrun{
-#'   # Download all metadata
-#'   metadata_all <- ifcb_download_dashboard_metadata("https://ifcb-data.whoi.edu/")
-#'
+#' \donttest{
 #'   # Download metadata for a specific dataset
-#'   metadata_mvco <- ifcb_download_dashboard_metadata("https://ifcb-data.whoi.edu/", "mvco")
+#'   metadata_mvco <- ifcb_download_dashboard_metadata("https://ifcb-data.whoi.edu/",
+#'                                                     dataset_name = "mvco",
+#'                                                     quiet = TRUE)
+#'
+#'   # Print result as tibble
+#'   dplyr::tibble(metadata_mvco)
 #' }
 #'
 #' @seealso [ifcb_download_dashboard_data()] to download data from the IFCB Dashboard API.